Refactor / organize CNF conversion code
