+++ src/gui.rs
@@ -12,7 +12,7 @@ use egui::Margin;
 use egui::RichText;
 
 use crate::{
-    app_state::AppState, cadical_wrapper::CadicalCallbackWrapper, cnf_var::CnfVariable,
+    app_state::AppState, cadical_wrapper::CadicalCallbackWrapper, cnf::CnfVariable,
     error::GenericError, gui::sudoku_cell::SudokuCell, ConstraintList, Trail,
 };
