+++ src/gui/sudoku_grid.rs
@@ -2,7 +2,7 @@ use std::cmp;
 
 use egui::{Color32, Pos2, Ui, Vec2};
 
-use crate::cnf_var::CnfVariable;
+use crate::cnf::CnfVariable;
 
 use super::SATApp;
