+++ src/lib.rs
@@ -1,23 +1,21 @@
 mod app_state;
-mod binary_cnf;
 mod cadical_wrapper;
-mod cnf_converter;
-mod cnf_var;
+mod cnf;
 mod error;
 mod filtering;
 pub mod gui;
+mod sudoku;
 
 #[cfg(test)]
 mod tests;
 
-use std::{cell::RefCell, fs, num::ParseIntError, path::Path, rc::Rc};
+use std::{cell::RefCell, num::ParseIntError, rc::Rc};
 
-use app_state::EncodingType;
 use cadical::Solver;
 
 use cadical_wrapper::CadicalCallbackWrapper;
-use cnf_converter::{clues_from_string, string_from_grid};
 use error::GenericError;
+use sudoku::{clues_from_string, string_from_grid};
 
 /// Rc<RefCell<Vec<Vec<i32>>>> is used to store the learned cnf_clauses
 #[derive(Clone)]
@@ -108,51 +106,6 @@ impl Default for Trail {
     }
 }
 
-pub fn solve_sudoku(
-    sudoku_clues: &[Vec<Option<i32>>],
-    solver: &mut Solver<CadicalCallbackWrapper>,
-    encoding: &EncodingType,
-) -> Result<Vec<Vec<Option<i32>>>, String> {
-    let mut solved: Vec<Vec<Option<i32>>> = Vec::new();
-    let cnf_clauses = encoding.sudoku_to_cnf(sudoku_clues);
-
-    for clause in cnf_clauses {
-        solver.add_clause(clause);
-    }
-
-    if let Some(true) = solver.solve() {
-        for row in 1..=9 {
-            let mut row_values = Vec::with_capacity(9);
-            for col in 1..=9 {
-                let value = encoding.get_cell_value(solver, row, col);
-                row_values.push(Some(value));
-            }
-            solved.push(row_values);
-        }
-        return Ok(solved);
-    }
-    Err(String::from("Solving sudoku failed!"))
-}
-
-pub fn get_sudoku(filename: String) -> Result<Vec<Vec<Option<i32>>>, GenericError> {
-    let sudoku_result = fs::read_to_string(filename);
-    match sudoku_result {
-        Ok(sudoku) => clues_from_string(sudoku, "."),
-        Err(_) => Err(GenericError {
-            msg: "Invalid filetype!".to_string(),
-        }),
-    }
-}
-
-pub fn write_sudoku(sudoku: String, path: &Path) -> Result<(), GenericError> {
-    let save_result = fs::write(path.display().to_string(), sudoku);
-    match save_result {
-        Err(_) => Err(GenericError {
-            msg: "Saving the file failed".to_string(),
-        }),
-        _ => Ok(()),
-    }
-}
 /// Parses the max_length filter input for applying the filter.
 pub fn parse_numeric_input(input: &str) -> Option<i32> {
     let parse_result: Result<i32, ParseIntError> = input.parse();
